reorganization of code related to output events
